+++ src/components/pill/pill.component.js
@@ -31,6 +31,7 @@ class Pill extends React.Component {
       fill,
       onDelete,
       colorVariant,
+      borderColor,
       pillRole,
       children,
       size,
@@ -47,6 +48,7 @@ class Pill extends React.Component {
         size={ size }
         ml={ ml }
         mr={ mr }
+        borderColor={ borderColor }
         { ...tagComponent('pill', this.props) }
       >
         { children }
@@ -57,30 +59,23 @@ class Pill extends React.Component {
 }
 
 Pill.propTypes = {
-
   /** Change the color of a status pill. */
   colorVariant: PropTypes.oneOf(['neutral', 'negative', 'positive', 'warning']),
-
+  /** Override color variant, provide any color from palette or any valid css color value. */
+  borderColor: PropTypes.string,
   /** The content to display inside of the pill.  */
   children: PropTypes.string.isRequired,
-
   /** Fills the pill background with colour. When fill is false only the border is coloured. */
   fill: PropTypes.bool,
-
   /** Sets the type of pill in use. */
   pillRole: PropTypes.oneOf(['tag', 'status']),
-
   /** Callback function for when the pill is clicked. */
   onClick: PropTypes.func,
-
   /** Callback function for when the remove icon is clicked. */
   onDelete: PropTypes.func,
-
   size: PropTypes.oneOf(['S', 'M', 'L', 'XL']),
-
   /** Margin right, given number will be multiplied by base spacing unit (8) */
   mr: PropTypes.number,
-
   /** Margin left, given number will be multiplied by base spacing unit (8) */
   ml: PropTypes.number
 };